Revert "Move atomic_ref and index_space_ids types into core from exp."
This reverts commit 59886c8b7bc32ad2c88a4027d5c5ea55d6573d6b.
