+++ numba_dpex/experimental/_kernel_dpcpp_spirv_overloads/_atomic_ref_overloads.py
@@ -25,7 +25,7 @@ from numba_dpex.kernel_api_impl.spirv.target import (
     LLVM_SPIRV_ARGS,
 )
 
-from ...core.types.kernel_api.atomic_ref import AtomicRefType
+from ..dpcpp_types import AtomicRefType
 from ..target import DPEX_KERNEL_EXP_TARGET_NAME
 from ._spv_atomic_inst_helper import (
     get_atomic_inst_name,
