+++ numba_dpex/experimental/core/types/__init__.py
@@ -1,0 +1,3 @@
+# SPDX-FileCopyrightText: 2024 Intel Corporation
+#
+# SPDX-License-Identifier: Apache-2.0
